# cedar-submission-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Better Metadata - Better Science\n\nThe submission server takes CEDAR metadata to the repositories that expect it in their own\nshape: NCBI BioSample and AIRR, ImmPort, and LINCS.\n\nTwo things happen here. Validation converts a CEDAR instance into the target format and returns\nthat repository's own verdict on it, sending nothing. Submission receives a package of files a\nchunk at a time, assembles it, and forwards it once the last chunk lands — so a successful\nresponse means the chunk was stored and any submission it completed was started, not that the\nrepository has accepted anything.\n",
    "title" : "CEDAR Submission Server API",
    "version" : "2.9.4"
  },
  "paths" : {
    "/command/immport-submit" : {
      "post" : {
        "description" : "Receive the files of an ImmPort submission, assemble them in CEDAR, and once they are all present submit them to the named workspace. The package arrives a chunk at a time, so a caller sends this repeatedly; the submission starts by itself once the last chunk lands. A 200 therefore means the chunk was stored and any submission it completed was started, not that the repository has accepted anything. While the package is still arriving the response is an empty object rather than a submission result.",
        "operationId" : "submitImmPort",
        "responses" : {
          "200" : {
            "description" : "The chunk was stored; ImmPort's result if this completed the submission, an empty object if not"
          },
          "400" : {
            "description" : "Not a multipart upload, or no workspace was named"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "500" : {
            "description" : "ImmPort could not be reached, or no ImmPort token is configured"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Upload a submission package and send it to ImmPort",
        "tags" : [ "ImmPort" ]
      }
    },
    "/command/immport-submit-old" : {
      "post" : {
        "description" : "The earlier ImmPort path, taking the whole submission as a single multipart request rather than in chunks. Kept for testing from the command line; the chunked route is what the workbench uses.",
        "operationId" : "submitImmPortOld",
        "responses" : {
          "200" : {
            "description" : "ImmPort's submission result"
          },
          "400" : {
            "description" : "Not a multipart upload, or no workspaceId parameter"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "500" : {
            "description" : "ImmPort could not be reached, or no ImmPort token is configured"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Send a submission to ImmPort in one request",
        "tags" : [ "ImmPort" ]
      }
    },
    "/command/immport-workspaces" : {
      "get" : {
        "description" : "Ask ImmPort which workspaces the configured account can submit to, and return them. A submission names one of these.",
        "operationId" : "immPortWorkspaces",
        "responses" : {
          "200" : {
            "description" : "The available ImmPort workspaces"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "500" : {
            "description" : "ImmPort could not be reached, or no ImmPort token is configured"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "List the caller's ImmPort workspaces",
        "tags" : [ "ImmPort" ]
      }
    },
...
  },
  "components" : {
    "securitySchemes" : {
      "api_key" : {
        "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}